requirements.txt
livekit-agents[openai,silero,turn-detector,deepgram,google,anthropic,cartesia,elevenlabs,rime,playai,groq,tavus,gladia,hedra,inworld]
livekit-plugins-aws[realtime]
livekit-plugins-noise-cancellation
python-dotenv
requests>=2.32.0
pydantic
flask
pandas
websockets>=11.0.3
rich
mcp